Create Inheritance
write a java program that demonstrates single inheritance. Create two classes in which one class contains the dimensions of the box with different types of contructor. and another class will inherit the base class and also contains the weight of box and finally call the mem. function into the main class.