Create abstract_class
What is the purpose of the abstract keyword in the A1 class of the given program in Java? How does the B1 class implement the abstract method print1()? Also, what happens when you uncomment the print2() method in the A1 class and try to call it on an object of B1 in the main() method?